+++ Changelog.md
@@ -1,3 +1,9 @@
+###26/01/2020
+
+- Tests / Pensum tests corrections; EmpFields class added for later use in Modules tests
+- Tests / Basic tests for Modules view
+- FIX / Pensum tests fixed (100% passing now)
+
 ###25/01/2020
 
 - Refactoring Tests